refactor(benchmarks): organize authored workload scenes
